// origin/src/youtube/types/ContinuedResults_0.ts
export interface ContinuedResults_0 {
    responseContext: ResponseContext
    contents: Contents
    trackingParams: string
    maxAgeStoreSeconds: number
    background: Background2
}

export interface ResponseContext {
    serviceTrackingParams: ServiceTrackingParam[]
    maxAgeSeconds: number
}

export interface ServiceTrackingParam {
    service: string
    params: Param[]
}

export interface Param {
    key: string
    value: string
}

export interface Contents {
    singleColumnBrowseResultsRenderer: SingleColumnBrowseResultsRenderer
}

export interface SingleColumnBrowseResultsRenderer {
    tabs: Tab[]
}
// ...
export interface Background2 {
    musicThumbnailRenderer: MusicThumbnailRenderer3
}

export interface MusicThumbnailRenderer3 {
    thumbnail: Thumbnail9
    thumbnailCrop: string
    thumbnailScale: string
    trackingParams: string
}